Conlusion of the research: no, it is not possible to create account with same Pubkey but different allocated space size.
Note: it is possible to recreate account (with same Pubkey) again if we set the same space size.

Transaction failes with 'custom program error 0x0' when trying to increase space for the same account. The lack of money to pay for the fees is unlikelly to be the cause of the transaction Error.

Note: by account I mean storage for state.